+++ gymnax/environments/classic_control/__init__.py
@@ -1,7 +1,4 @@
-from gymnax.environments.classic_control.pendulum import (reset_pendulum,
-                                                          step_pendulum,
-                                                          params_pendulum)
-
+from gymnax.environments.classic_control.pendulum import Pendulum
 from gymnax.environments.classic_control.cartpole import CartPole
 
 from gymnax.environments.classic_control.mountain_car import (reset_mountain_car,