+++ CHANGELOG.md
@@ -1,1 +1,15 @@
-**v0.0.1**
+### v0.0.1 - Unreleased
+
+##### Added
+- Adds main infrastructure:
+    - Jittable environment base class
+    - Spaces: `Discrete`, `Continuous`, `Box`
+- Adds base set of environments:
+    - OpenAI's `classic_control`
+    - DeepMind's `bsuite`
+    - `MinAtar`
+
+##### Changed
+
+
+##### Fixed